.github/scripts/requirements.txt
markdown_to_json==2.1.2
tabulate==0.9.0
pytz==2024.1
pydantic==2.11.1
pydantic_core==2.33.0
google-cloud-aiplatform==1.106.0
vertexai==1.43.0
PyGithub==2.7.0
pyyaml==6.0.2